feat: show similar keys first on KeyInFile error
